Update dep github.com/skeema/tengo
This brings in a temporary fix to allow different default params between a
DockerizedInstance (for testing) and its corresponding usage in Skeema's tests.

This will later be removed once caching logic in Tengo is eliminated, but that
will be a substantial refactor which would benefit from having test coverage
first.